Refactor DIContainer
