+++ BuyPolish/Pola/Manager/BarcodeValidator/BarcodeValidator.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol BarcodeValidator {
+    func isValid(barcode: String) -> Bool
+}
