+++ BuyPolish/Pola/Manager/BarcodeValidator/BarcodeValidatorRegistrant.swift
@@ -1,0 +1,10 @@
+import Foundation
+import Swinject
+
+final class BarcodeValidatorRegistrant: DependencyRegistrant {
+    func registerDependency(container: Container) {
+        container.register(BarcodeValidator.self) { _ in
+            EANBarcodeValidator()
+        }
+    }
+}
