+++ BuyPolish/Pola/Manager/CapturedImageManager/CapturedImageManagerRegistrant.swift
@@ -1,0 +1,9 @@
+import Swinject
+
+final class CapturedImageManagerRegistrant: DependencyRegistrant {
+    func registerDependency(container: Container) {
+        container.register(CapturedImageManager.self) { _ in
+            InMemoryCapturedImageManager()
+        }
+    }
+}
