+++ BuyPolish/Pola/Manager/CodeScannerManager/CodeScannerManagerRegistrant.swift
@@ -1,0 +1,10 @@
+import Foundation
+import Swinject
+
+final class CodeScannerManagerRegistrant: DependencyRegistrant {
+    func registerDependency(container: Container) {
+        container.register(CodeScannerManager.self) { _ in
+            CameraSessionCodeScannerManager()
+        }
+    }
+}
