+++ BuyPolish/Pola/Manager/DI/DependencyRegistrant.swift
@@ -1,0 +1,5 @@
+import Swinject
+
+protocol DependencyRegistrant {
+    func registerDependency(container: Container)
+}
