+++ BuyPolish/Pola/Manager/FlashlightManager/FlashlightManager.swift
