+++ BuyPolish/Pola/Manager/FlashlightManager/FlashlightManagerRegistrant.swift
@@ -1,0 +1,10 @@
+import Foundation
+import Swinject
+
+final class FlashlightManagerRegistrant: DependencyRegistrant {
+    func registerDependency(container: Container) {
+        container.register(FlashlightManager.self) { _ in
+            FlashlightManager()
+        }
+    }
+}
