+++ BuyPolish/Pola/Manager/Logger/LoggerRegistrant.swift
@@ -6,11 +6,11 @@ final class LoggerRegistrant: DependencyRegistrant {
         if FirebaseApp.isFirebaseAvailable {
             container.register(LoggerProvider.self) { _ in
                 CrashlitycsLogger()
-            }
+            }.inObjectScope(.container)
         } else {
             container.register(LoggerProvider.self) { _ in
                 ConsoleLogger()
-            }
+            }.inObjectScope(.container)
         }
     }
 }
