+++ BuyPolish/Pola/Manager/ProductImageManager/ProductImageManagerRegistrant.swift
@@ -1,0 +1,9 @@
+import Swinject
+
+final class ProductImageManagerRegistrant: DependencyRegistrant {
+    func registerDependency(container: Container) {
+        container.register(ProductImageManager.self) { resolver in
+            LocalDocumentsProductImageManager(fileManager: resolver.resolve(FileManager.self)!)
+        }
+    }
+}
